Directive Assignment
Assignment: Make a button that toggles an element. This button should also add items to an array. This array should then be displayed using ngFor. Finally if the index of an item is greater or equal to 5 then we need to style it with a blue background, and give it a class for white text.